ASSIGNED - task 242: WTX.remove is not adapting the childcount in a correct way https://projects.uni-konstanz.de/issues/show/242
removed rtx.gerRevisionSize since it cannot be adapted in a scaling way. The entire subtree has to be counted when a remove occurs. Of coursee, this can be done in a multithreaded way and is an easy query. However, no mandatory parts of the application of this value so I removed it entirely.

However, the TreeTankDebugger is lost as well...who needs debugging?

git-svn-id: https://svn.uni-konstanz.de/disy/Treetank/trunk/treetank@5146 e3ddb328-5bfe-0310-b762-aafcbcbd2528