ASSIGNED - task 1406: Node + Nodepage refactoring https://projects.uni-konstanz.de/issues/show/1406
Removed TypeValue-calls sind they are not needed and are constricting delegation-methods of node-interfaces and Delegation classes.
Furthermore, IItem was moved and renamed to INode and new packages in the node-package for interfaces and delegation classes are introduced.
All testcases are fine excluding FunctionsTest which are ignored but marked by Ignores.

git-svn-id: https://svn.uni-konstanz.de/disy/Treetank/trunk/treetank@6405 e3ddb328-5bfe-0310-b762-aafcbcbd2528